code generation for global variables

+++ test/c1.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
+#include <math.h>
 
-int v = 123;
+short v = 123;
 
 int main(int argc, char** args) {
     printf( "%d\n", v );
